Fix isFacetedSearchEnabled scope for the correct pagelayout
The scope level of the isFacetedSearchEnabled version is global which makes no sense since all config fields have to be configured on store level. Whether the value is true or false, it will always return false and set the search result page to a 1column page (even when the faceted search functionality is enabled).